services/ai-agents/pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "FastAPI service for AI agents and model orchestration."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn[standard]>=0.29.0,<0.31.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pydantic>=2.7.0,<3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
